⚡️ Microoptimize SemverPreRelease ToString call
Apparently, `UInt32.ToString()` call requires a variable's address, and because the value of `number` is cast to `uint32` before the call, the value can't be accessed directly, so the compiler creates a `uint32` variable. Luckily, no other places are affected by this change, since, aside from this, `int32` is essentially equivalent to `uint32`.